+++ Changelog.md
@@ -29,6 +29,7 @@ Enhancements
 * Times in profile output are now bold instead of failure_color.
   (Matthew Boedicker)
 * Add --no-fail-fast command line option. (Gonzalo Rodríguez-Baltanás Díaz)
+* Clear shared example registry using accessor. (thepoho, Jon Rowe)
 
 ### 2.14.4 / 2013-07-21
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.14.3...v2.14.4)